% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipoPieChart.R
\name{lipoPieChart}
\alias{lipoPieChart}
\title{lipoPieChart
Pie charts and tables for lipoproteins.}
\usage{
lipoPieChart(
  data,
  group,
  subfractions = T,
  subcompositions = T,
  cohort = "1",
  optns = list()
)
}
\arguments{
\item{data}{A data frame of the lipoproteins.}

\item{group}{A character vector of the groups the same length as the data,
for example Female_HighBMI, Female_LowBMI, Male_HighBMI, Male_LowBMI.}

\item{subfractions}{Logical for if subfractions of HDL, VLDL and LDL should
be produced. Default is TRUE. looks at subfractions considering the lipid
distribution, so what portion do each of V1TG, V2TG, V3TG, V4TG, V5TG
contribute to VLTG}

\item{subcompositions}{Logical for if subcompositions of HDL, VLDL and LDL
should be produced. Looks at subfractions considering the lipid composition,
so what portion does each of the raw V1TG, V1CH, V1PL contribute to the
calculated V1TL.}

\item{cohort}{A character vector of the cohorts the same length as the data.
Used to facet rows of the pie charts. If only one cohort is used, can supply
just the name of the cohort as a character "AUS" for example. If no cohort
label desired leave as default, 1, and no label will appear.}

\item{optns}{empty list for future non-essential arguments.}
}
\description{
lipoPieChart
Pie charts and tables for lipoproteins.
}
