% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva.methods-package.R
\docType{package}
\name{uva.plots-package}
\alias{uva.plots}
\alias{uva.plots-package}
\title{uva.plots: plot functions for univariate models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
